Fix aria-labelledby to listen to store
