+++ src/app/components/Results/Results.jsx
@@ -21,6 +21,7 @@ import ReturnLink from '../ReturnLink/ReturnLink.jsx';
 // Import utilities
 import { makeClientApiCall } from '../../utils/MakeClientApiCall.js';
 import { generateSearchedFrom, nativeGA } from '../../utils/GAUtils.js';
+import getNumberForFacet from '../../utils/TabIndex.js'
 
 class Results extends React.Component {
   constructor(props) {
@@ -342,7 +343,7 @@ class Results extends React.Component {
               viewBox="0 0 84 4"
               width="84"
             />
-            <ol id={this.props.id} className={this.props.className} ref="resultsOlElement" tabIndex="0" aria-labelledby={`link${this.state.tabIdValue}`}>
+          <ol id={this.props.id} className={this.props.className} ref="resultsOlElement" tabIndex="0" aria-labelledby={`link${getNumberForFacet(this.props.selectedFacet)}`}>
               {results}
             </ol>
             {
